requirements.txt
fastapi
uvicorn
python-dotenv
openai
discord.py
aiohttp
